@@ -1,74 +1,95 @@
 {
-    "name": "medcenter24/mc-core",
-    "description": "Backend for the MedCenter24 project",
-    "keywords": ["api", "medcenter24", "backoffice", "medicine"],
-    "license": "GPL-2.0-only",
-    "minimum-stability": "dev",
-    "type": "medcenter24-extension",
-    "prefer-stable": true,
-    "extra": {
-        "medcenter24-extension-name": "mcCore",
-        "laravel": {
-            "dont-discover": [
-                "cmgmyr/messenger"
-            ]
-        }
-    },
-    "require": {
-        "php": ">=8.0",
-        "ext-dom": "*",
-        "ext-json": "*",
-        "ext-libxml": "*",
-        "cmgmyr/messenger": "^2.16.2",
-        "dingo/api": "~3.0",
-        "fideloper/proxy": "^4.4",
-        "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^7.3.0",
-        "intervention/image": "^2.6",
-        "irazasyed/telegram-bot-sdk": "^v3.4.1",
-        "laravel/framework": "^8.47",
-        "laravel/tinker": "~2.0",
-        "laravel/ui": "^3.3",
-        "maatwebsite/excel": "^3.1.19",
-        "medcenter24/lib-installer": "^2.0.1",
-        "mpdf/mpdf": "~8.0.13",
-        "nwidart/laravel-modules": "^8.2",
-        "spatie/laravel-medialibrary": "^9.6",
-        "srmklive/flysystem-dropbox-v2": "^v1.0.8",
-        "tymon/jwt-auth": "^1.0.2"
-    },
-    "require-dev": {
-        "barryvdh/laravel-debugbar": "^3.6",
-        "barryvdh/laravel-ide-helper": "^2.10",
-        "filp/whoops": "^2.7.1",
-        "facade/ignition": "^2.8.3",
-        "fzaninotto/faker": "^1.9",
-        "mockery/mockery": "^1.4",
-        "nunomaduro/collision": "^5.5",
-        "phpunit/phpunit": "^9.5",
-        "roave/security-advisories": "dev-master"
-    },
-    "autoload": {
-        "psr-4": {
-            "medcenter24\\mcCore\\App\\": "app/",
-            "Database\\Factories\\": "database/factories/",
-            "Database\\Seeders\\": "database/seeds/"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "medcenter24\\mcCore\\Tests\\": "Tests"
-        }
-    },
-    "scripts": {
-        "post-autoload-dump": [
-            "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
-            "@php artisan package:discover --ansi"
-        ]
-    },
-    "config": {
-        "optimize-autoloader": true,
-        "preferred-install": "dist",
-        "sort-packages": true
+  "name": "medcenter24/mc-core",
+  "description": "Backend for the MedCenter24 project",
+  "keywords": [
+    "api",
+    "medcenter24",
+    "backoffice",
+    "medicine"
+  ],
+  "license": "GPL-2.0-only",
+  "minimum-stability": "dev",
+  "type": "medcenter24-extension",
+  "prefer-stable": true,
+  "extra": {
+    "medcenter24-extension-name": "mcCore",
+    "laravel": {
+      "dont-discover": [
+        "cmgmyr/messenger"
+      ]
     }
+  },
+  "require": {
+    "php": ">=8.1",
+    "ext-dom": "*",
+    "ext-json": "*",
+    "ext-libxml": "*",
+    "cmgmyr/messenger": "^2.27.0",
+    "api-ecosystem-for-laravel/dingo-api": "dev-laravel-9-support",
+    "doctrine/dbal": "^3.3",
+    "fruitcake/laravel-cors": "^2.0",
+    "guzzlehttp/guzzle": "^7.4.1",
+    "intervention/image": "^2.7.1",
+    "irazasyed/telegram-bot-sdk": "^3.4.1",
+    "laravel/framework": "^9.4",
+    "laravel/tinker": "~2.7",
+    "laravel/ui": "^3.4",
+    "maatwebsite/excel": "^3.1.37",
+    "medcenter24/lib-installer": "^2.0.1",
+    "mpdf/mpdf": "~v8.0.17",
+    "nwidart/laravel-modules": "^9.0.1",
+    "spatie/laravel-medialibrary": "^10.1.3",
+    "php-open-source-saver/jwt-auth": "^1.4.1"
+  },
+  "require-dev": {
+    "barryvdh/laravel-debugbar": "^3.6",
+    "barryvdh/laravel-ide-helper": "^v2.12.3",
+    "filp/whoops": "^2.14.3",
+    "spatie/laravel-ignition": "^1.0.7",
+    "fzaninotto/faker": "^1.9",
+    "mockery/mockery": "^1.5.0",
+    "nunomaduro/collision": "^v6.1.0",
+    "phpunit/phpunit": "^9.5",
+    "roave/security-advisories": "dev-master"
+  },
+  "autoload": {
+    "psr-4": {
+      "medcenter24\\mcCore\\App\\": "app/",
+      "Database\\Factories\\": "database/factories/",
+      "Database\\Seeders\\": "database/seeds/"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "medcenter24\\mcCore\\Tests\\": "Tests"
+    }
+  },
+  "scripts": {
+    "post-autoload-dump": [
+      "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
+      "@php artisan package:discover --ansi"
+    ]
+  },
+  "config": {
+    "optimize-autoloader": true,
+    "preferred-install": "dist",
+    "sort-packages": true,
+    "allow-plugins": {
+      "medcenter24/lib-installer": true
+    }
+  },
+  "repositories": [
+    {
+      "type": "vcs",
+      "url":  "git@github.com:christoph-kluge/dingo-api.git"
+    },
+    {
+      "type": "vcs",
+      "url":  "git@github.com:dmason30/blueprint.git"
+    },
+    {
+      "type": "vcs",
+      "url": "https://github.com/annuh/fractal.git"
+    }
+  ]
 }
