@@ -42,7 +42,7 @@ class AccessTest extends TestCase
     public function testAdmin(): void
     {
         Roles::shouldReceive('hasRole')
-            ->times(9)
+            ->times(1)
             ->andReturnUsing(static function ($user, $role) {
                 return true;
             });
@@ -51,13 +51,14 @@ class AccessTest extends TestCase
             ->actingAs(User::factory()->make())
             ->get('/admin');
 
-        $response->assertStatus(200);
+        $response->assertStatus(302);
+        $response->assertRedirect('/admin/users');
     }
 
     public function testDoctor(): void
     {
         Roles::shouldReceive('hasRole')
-            ->times(5)
+            ->times(1)
             ->andReturnUsing(function ($user, $role) {
                 return false;
             });
@@ -72,7 +73,7 @@ class AccessTest extends TestCase
     public function testDirector(): void
     {
         Roles::shouldReceive('hasRole')
-            ->times(5)
+            ->times(1)
             ->andReturnUsing(function ($user, $role) {
                 return false;
             });
