@@ -48,16 +48,16 @@ class AccidentServiceTest extends TestCase
 
         $this->assertSame([
             'id' => 1,
-            'parent_id' => '0',
-            'patient_id' => '0',
-            'accident_type_id' => '0',
-            'accident_status_id' => '1',
-            'assistant_id' => '0',
+            'parent_id' => 0,
+            'patient_id' => 0,
+            'accident_type_id' => 0,
+            'accident_status_id' => 1,
+            'assistant_id' => 0,
             'assistant_ref_num' => '',
-            'assistant_invoice_id' => '0',
-            'assistant_guarantee_id' => '0',
-            'form_report_id' => '0',
-            'city_id' => '0',
+            'assistant_invoice_id' => 0,
+            'assistant_guarantee_id' => 0,
+            'form_report_id' => 0,
+            'city_id' => 0,
             'ref_num' => '',
             'title' => '',
             'address' => '',
