@@ -90,14 +90,14 @@ class DatePeriodServiceTest extends TestCase
         self::assertSame($data, $datePeriod->toArray());
         self::assertSame([[
             'id' => 1,
-            'date_period_id' => '1',
-            'day_of_week' => '0',
+            'date_period_id' => 1,
+            'day_of_week' => 0,
             'from' => '12:21',
             'to' => '23:59',
         ], [
             'id' => 2,
-            'date_period_id' => '1',
-            'day_of_week' => '0',
+            'date_period_id' => 1,
+            'day_of_week' => 0,
             'from' => '00:00',
             'to' => '22:15',
         ]], $datePeriod->interpretation->toArray());
