@@ -40,7 +40,7 @@ class CaseCaseableController extends ApiController
     }
 
     /**
-     * @param $id
+     * @param int $id
      * @return Response
      */
     public function getDoctorCase(int $id): Response
@@ -60,7 +60,7 @@ class CaseCaseableController extends ApiController
     }
 
     /**
-     * @param $id
+     * @param int $id
      * @return Response
      */
     public function getHospitalCase(int $id): Response
