@@ -135,9 +135,9 @@ class AccidentService extends AbstractModelService
     ];
 
     public const DATE_FIELDS = [
-        AccidentService::FIELD_CREATED_AT,
-        AccidentService::FIELD_DELETED_AT,
-        AccidentService::FIELD_UPDATED_AT,
+        AbstractModelService::FIELD_CREATED_AT,
+        AbstractModelService::FIELD_DELETED_AT,
+        AbstractModelService::FIELD_UPDATED_AT,
         AccidentService::FIELD_HANDLING_TIME,
     ];
 
@@ -302,7 +302,7 @@ class AccidentService extends AbstractModelService
      * @param string $comment
      * @throws InconsistentDataException
      */
-    public function rejectDoctorAccident($accident, $comment = 'rejected'): void
+    public function rejectDoctorAccident($accident, string $comment = 'rejected'): void
     {
         $status = $this->getAccidentStatusesService()->getDoctorRejectedStatus();
         $this->setStatus($accident, $status, $comment);
@@ -313,16 +313,26 @@ class AccidentService extends AbstractModelService
      * @param string $comment
      * @throws InconsistentDataException
      */
-    public function closeAccident(Accident $accident, $comment = 'closed'): void
+    public function closeAccident(Accident $accident, string $comment = 'closed'): void
     {
         $this->setStatus($accident, $this->getAccidentStatusesService()->getClosedStatus(), $comment);
     }
 
     /**
-     * @param array $data
-     * @return Model
+     * @param Accident $accident
+     * @param string $comment
+     * @throws InconsistentDataException
      */
-    public function create(array $data = []): Model
+    public function reopenAccident(Accident $accident, string $comment = 'reopened'): void
+    {
+        $this->setStatus($accident, $this->getAccidentStatusesService()->getReopenedStatus(), $comment);
+    }
+
+    /**
+     * @param array $data
+     * @return Model|Accident
+     */
+    public function create(array $data = []): Model|Accident
     {
         /** @var Accident $accident */
         $accident = parent::create($data);
@@ -344,7 +354,7 @@ class AccidentService extends AbstractModelService
             $this->convertToFilter($filterByFields, $data)
         );
 
-        if ($this->isClosed($previousAccident)) {
+        if ($this->isClosed($previousAccident) && !$this->isReopening($data)) {
             throw new InconsistentDataException('Accident closed and can not be changed', 422);
         }
 
@@ -362,5 +372,13 @@ class AccidentService extends AbstractModelService
     public function isHospitalAccident(Accident $accident): bool
     {
         return $accident->getAttribute(self::FIELD_CASEABLE_TYPE) === HospitalAccident::class;
+    }
+
+    private function isReopening(array $data): bool
+    {
+        $reopenedStatus = $this->getAccidentStatusesService()->getReopenedStatus();
+        return !empty($data)
+            && ($data[self::FIELD_ACCIDENT_STATUS_ID] ?? null)
+            === $reopenedStatus->getAttribute(self::FIELD_ID);
     }
 }
