@@ -16,6 +16,8 @@
  * Copyright (c) 2019 (original work) MedCenter24.com;
  */
 
+use Illuminate\Support\Str;
+
 return [
 
     /*
@@ -102,6 +104,5 @@ return [
     |
     */
 
-    'prefix' => 'laravel',
-
+    'prefix' => env('CACHE_PREFIX', Str::slug(env('APP_NAME', 'laravel'), '_').'_cache_'),
 ];
