@@ -52,12 +52,12 @@ return [
     /*
      * Sets the Access-Control-Expose-Headers response header.
      */
-    'exposed_headers' => false,
+    'exposed_headers' => [],
 
     /*
      * Sets the Access-Control-Max-Age response header.
      */
-    'max_age' => false,
+    'max_age' => 0,
 
     /*
      * Sets the Access-Control-Allow-Credentials header.
