@@ -18,8 +18,8 @@
 
 return [
     'welcome_guest' => 'Вас приветствует Бот медицинской компании.',
-    'welcome_user' => 'Привет :user, с возвращением. Я ждал тебя.',
-    'invite_request' => 'Мы не узнали Вас, для использования кода введите Код-приглашение.',
+    'welcome_user' => 'Привет :user, с возвращением.',
+    'invite_request' => 'Мы не узнали Вас, для подключения аккаунта к боту введите Код-приглашение.',
     'incorrect_invite_token' => 'Ваш код не прошел валидацию. Попробуйте еще раз: используйте комманду /start чтобы попробовать еще раз.',
     'invite_success' => 'Вы подключили свой аккаунт к Telegram. Используйте комманду /help для получения списка доступных комманд.',
     'command_start_description' => 'Начало использования сервиса',
